+++ Afvaldienst/__init__.py
@@ -3,6 +3,6 @@ __author_email__ = 'spam@rootrulez.com'
 __license__ = 'MIT'
 __maintainer_email__ = 'spam@rootrulez.com'
 __url__ = 'https://github.com/xirixiz/afvaldienst',
-__version__ = '1.0.3'
+__version__ = '1.0.5'
 
 from .Afvaldienst import Afvaldienst